Create a git hook for commit messages
Add the git hook git-good-commit that examines commit messages. It
validates commit messages conform to "the seven rules of a great git
commit message". Additionally, It provides a prompt if any of these
guidelines are broken.

To properly configure the git hooks, the following commands should be
run each time the repository is cloned:
$ make init