Fix image links in doxygen website
To fix the broken link in the README.md file on the Doxygen website,
move README.md and CODE OF CONDUCT.md to the folder where the images are
placed. And to make the contents better fit the documentation, update
the links in README.md and the rule of cleaning in Makefile.